Update annotators_summary field type to BIGINT in MetadataStore schema

+++ databricks-unity-dxr-integration/src/databricks_unity_dxr_integration/metadata_store.py
@@ -93,7 +93,7 @@ class MetadataStore:
                 connector_type STRING,
                 connector_site_url STRING,
                 annotators_json STRING,
-                annotators_summary MAP<STRING, INT>,
+                annotators_summary MAP<STRING, BIGINT>,
                 collected_at TIMESTAMP
             )
             USING DELTA
